Moved tests to same directories as tested files
