+++ src/components/businesses/BusinessModals.test.js
@@ -1,6 +1,6 @@
 import { mount } from 'enzyme';
 import React from 'react'
-import BusinessModals from '../../components/businesses/BusinessModals';
+import BusinessModals from '../BusinessModals';
 
 describe('<BusinessesList />', () => {
 	it('it renders correctly', async () => {
